requirements.txt
Flask
Flask-Cors
PyMuPDF
openai
python-dotenv
gunicorn